refactor(user): refactor create meeting modal for paginated forms
